==> 0_etl-AvazuCTR.ipynb <==
first_time = True
!ls

if first_time:
  from google.colab import drive
  drive.mount("/content/drive")

if first_time:
  !apt-get update
  !apt-get install openjdk-8-jdk-headless -qq > /dev/null
  !wget -q https://archive.apache.org/dist/spark/spark-3.0.0/spark-3.0.0-bin-hadoop2.7.tgz
  !tar xf spark-3.0.0-bin-hadoop2.7.tgz

# Set up Spark
!pip install -q findspark
!pip install py4j

!export JAVA_HOME=$(/usr/lib/jvm/java-8-openjdk-amd64 -v 1.8)
! echo $JAVA_HOME
import os
os.environ["JAVA_HOME"] = "/usr/lib/jvm/java-8-openjdk-amd64"
os.environ["SPARK_HOME"] = "/content/spark-3.0.0-bin-hadoop2.7"
import findspark
findspark.init("spark-3.0.0-bin-hadoop2.7")# SPARK_HOME

from pyspark.sql import SparkSession
spark = SparkSession.builder.master("local[*]").getOrCreate()

if first_time:
  !pip install koalas
  !pip install --upgrade PyArrow

import time
import databricks.koalas as ks
import numpy as np
import pandas as pd
from IPython.display import display # for displaying multiple  dataframes from a single cell
import seaborn as sns
import matplotlib.pyplot as plt
import math
from numpy.linalg import norm
import os
os.environ["PYSPARK_PYTHON"] = "python3"

import os
os.environ["PYARROW_IGNORE_TIMEZONE"] = "1"

%matplotlib inline
from pyspark.sql import SparkSession
spark = SparkSession \
    .builder \
    .appName("AvazuCTR") \
    .config("spark.some.config.option", "some-value") \
    .getOrCreate()

## load data

!cp /content/drive/MyDrive/LAI_OFFER/BIG_DATA/AvazuCTR/Data/ARCHIVED/train.csv .

df = spark.read.csv("train.csv",inferSchema=True,header=True)

df.show(10)

## count

df.count()

## Schema

df.printSchema()

# Save data in parquet

Parquet has some advantages over csv, primarily ([source](https://www.youtube.com/results?search_query=spark++Parquet)):
- defined schemas
- predicate push down, e.g., filtering before processing

'''
df.coalesce(4) #Returns a new DataFrame that has exactly numPartitions partitions. 
  .write #write a DataFrame to external storage systems (e.g. file systems, key-value stores, etc). Use DataFrame.write to access this.
  .mode("overwrite") #Specifies the behavior when data or table already exists.
  .parquet("parquet/filtered_train.csv")# Parquet storage formats is more efficient/fast with big data
# speed up sql querel by overcoming two major bottlenecks of distributed analytics: communication costs (IO bound) and data decoding (CPU bound).
'''
df.coalesce(4).write.mode('overwrite').parquet("/content/drive/MyDrive/LAI_OFFER/BIG_DATA/AvazuCTR/Data/parquet/train.csv")

coalesce is like repartition with minimal shuffling and data moval 

Complementary reading:
- coalesce is like repartition with minimal shuffling and data moval 
- Spark can run 1 concurrent task for every partition of an RDD (up to the number of cores in the cluster). If you're cluster has 20 cores, you should have at least 20 partitions (in practice 2–3x times more ) [ref](https://www.google.com/search?q=spark+number+of+partitions&oq=spark+number+of+partitions&aqs=chrome..69i57j0i7i30l7.9744j0j7&sourceid=chrome&ie=UTF-8)
- [partition-worker node,   excutor-task](http://beginnershadoop.com/2019/09/30/distribution-of-executors-cores-and-memory-for-a-spark-application/)
- [stackoverflow](https://stackoverflow.com/questions/24696777/what-is-the-relationship-between-workers-worker-instances-and-executors#:~:text=A%20node%20is%20a%20machine,many%20executors%2C%20for%20many%20applications.)

- refer to class 59
- excutor: a single JVM INSTANCE(/process?) ON A NODETHAT SERVERS A SINGLE SPARK application. An executor runs multiple tasks over its lifetime, and multiple tasks concurrently (?). A node may have several Spark executors and there are many nodes running Spark Executors for each client application.
- task: spark task represents a unit of work on a partition of a distributed dataset (RDD). 

An executor is a single JVM process which is launched for an application on a worker node. excutors are for running tasks, A single node can run multiple executors and executors for an application can span multiple worker nodes. An executor stays up for the duration of the Spark Application and runs the tasks in multiple threads. 
- core: A core is a basic computation unit of CPU and a CPU may have one or more cores to perform tasks at a given time. The more cores we have, the more work we can do. In spark, this controls the number of parallel tasks an executor can run. Liu: It seems cores can be shared by many excutors in the same single worker node.
- Workers hold many executors, for many applications. One application could have executors on many workers. In Colab, only a single application is lauched.
