==> 2_make_petastorm_loader_wDash.ipynb <==
first_time = True
full_data = True

copy_full_data = True
copy_sample_data = False

if first_time:
  from google.colab import drive
  drive.mount("/content/drive")

if first_time:
  !pip install petastorm[torch]

if first_time:
  !apt-get update
  !apt-get install openjdk-8-jdk-headless -qq > /dev/null
  # !wget -q https://downloads.apache.org/spark/spark-2.4.7/spark-2.4.7-bin-hadoop2.7.tgz
  # !tar xf spark-2.4.7-bin-hadoop2.7.tgz
  !wget -q https://archive.apache.org/dist/spark/spark-3.0.0/spark-3.0.0-bin-hadoop2.7.tgz
  !tar xf spark-3.0.0-bin-hadoop2.7.tgz

# Set up Spark
if first_time:
  !pip install -q findspark
  !pip install py4j

!export JAVA_HOME=$(/usr/lib/jvm/java-8-openjdk-amd64 -v 1.8)
! echo $JAVA_HOME
import os
os.environ["JAVA_HOME"] = "/usr/lib/jvm/java-8-openjdk-amd64"
# os.environ["SPARK_HOME"] = "/content/spark-2.4.7-bin-hadoop2.7"
os.environ["SPARK_HOME"] = "/content/spark-3.0.0-bin-hadoop2.7"

import findspark
# findspark.init("spark-2.4.7-bin-hadoop2.7")# SPARK_HOME
findspark.init("spark-3.0.0-bin-hadoop2.7")# SPARK_HOME
if first_time:
  !pip install --upgrade PyArrow
import os
os.environ["PYSPARK_PYTHON"] = "python3"
os.environ["PYARROW_IGNORE_TIMEZONE"] = "1"


from pyspark.sql import SparkSession

  spark = SparkSession \
      .builder \
      .appName("AvazuCTR") \
      .config("spark.some.config.option", "some-value") \
      .getOrCreate()

## Load processed data

# copy parquet files from Google Drive
if copy_full_data:
  !cp -R /content/drive/MyDrive/LAI_OFFER/BIG_DATA/AvazuCTR/WDL_FTRL/data_wDash/train_df_parquet .
  !cp -R /content/drive/MyDrive/LAI_OFFER/BIG_DATA/AvazuCTR/WDL_FTRL/data_wDash/test_df_parquet .
if copy_sample_data:
  !cp -R /content/drive/MyDrive/LAI_OFFER/BIG_DATA/AvazuCTR/WDL_FTRL/data_wDash/train_df_parquet_filtered .
  !cp -R /content/drive/MyDrive/LAI_OFFER/BIG_DATA/AvazuCTR/WDL_FTRL/data_wDash/test_df_parquet_filtered .
  
# a larger number of partitions makes a partition smaller than the allocated excutor memory ensuring performance
# partition usually will be different fro where saved
# https://stackoverflow.com/questions/50825835/does-spark-maintain-parquet-partitioning-on-read
if full_data:
  train_df = spark.read.parquet("train_df_parquet")#.repartition(4) #repartition(64)
  test_df = spark.read.parquet("test_df_parquet")
else:
  train_df = spark.read.parquet("train_df_parquet_filtered")#.repartition(4) #repartition(64)
  test_df = spark.read.parquet("test_df_parquet_filtered")


# repartition by 'date', lazy
from pyspark.sql.functions import col as column
train_df = train_df.repartition(9,column('date'))
test_df = test_df.repartition(9,column('date'))

# data type conversion for 'id'
from pyspark.sql.types import FloatType,IntegerType,DoubleType
train_df = train_df.withColumn('id',  column('id').cast(DoubleType()))
test_df = test_df.withColumn('id',  column('id').cast(DoubleType()))
train_df.printSchema()

# sort by 'date', lazy, this ordering will be messed up by petastorm
# train_df = train_df.orderBy('date',ascending = True)
# test_df = test_df.orderBy('date',ascending = True)

import pickle
path_processed_features = '/content/drive/MyDrive/LAI_OFFER/BIG_DATA/AvazuCTR/WDL_FTRL/data_wDash/1_SparkProcess_AvazuCTR.pickle'
with open(path_processed_features,'rb') as handle:  
  wide_cols = pickle.load(handle)
  interactive_features = pickle.load(handle)
  D = pickle.load(handle)
  embeddings_cols = pickle.load(handle)
  loader_cols = pickle.load(handle)
print('len wide_cols',len(wide_cols))
print('len interactive_features',len(interactive_features))
print('D',D)
print('len embeddings_cols',len(embeddings_cols))
print('len(loader_cols)',len(loader_cols))

print('-----------')

print('- wide_cols',wide_cols)
print('- interactive_features',interactive_features)
print('- D',D)
print('- embeddings_cols',embeddings_cols)
print('- loader_cols',loader_cols)

=============

## make petastorm intemediate data

from petastorm.spark import SparkDatasetConverter, make_spark_converter
from petastorm import TransformSpec

!rm -r /content/petastorm # remove files so as to make everything inside petastorm are generated by current Python session
# set a cache directory for intermediate data, hold by "SparkDatasetConverter" object
spark.conf.set(SparkDatasetConverter.PARENT_CACHE_DIR_URL_CONF, "file:///content/petastorm/cache")
converter_train = make_spark_converter(train_df,dtype = 'float32') # int features will be kept as int
converter_test = make_spark_converter(test_df,dtype = 'float32')

print(len(converter_train))
print(len(converter_test))
print(converter_train.cache_dir_url)
print(converter_test.cache_dir_url)

## save data

# copy cached files to google drive
if full_data:
  !cp -r /content/petastorm    /content/drive/MyDrive/LAI_OFFER/BIG_DATA/AvazuCTR/WDL_FTRL/data_wDash/petastorm
else:
  !cp -r /content/petastorm   /content/drive/MyDrive/LAI_OFFER/BIG_DATA/AvazuCTR/WDL_FTRL/data_wDash/petastorm_filtered

# copy converter object to google drive
file_name  = 'converter'
if full_data:
  converter_path = F"/content/drive/My Drive/LAI_OFFER/BIG_DATA/AvazuCTR/WDL_FTRL/data_wDash/petastorm/{file_name}.pickle"
else:
  converter_path = F"/content/drive/My Drive/LAI_OFFER/BIG_DATA/AvazuCTR/WDL_FTRL/data_wDash/petastorm_filtered/{file_name}.pickle"


# save 'converter_features_label' object (serialize)
import pickle
with open(converter_path, 'wb') as handle:
    pickle.dump(converter_train, handle, protocol=pickle.HIGHEST_PROTOCOL)
    pickle.dump(converter_test, handle, protocol=pickle.HIGHEST_PROTOCOL)

## make data loader and test it

def get_transform_spec():
  return TransformSpec(func = None, selected_fields = loader_cols)

batch_size = 32
for epoch in range(1):
  with converter_train.make_torch_dataloader(batch_size = batch_size, transform_spec = get_transform_spec(),num_epochs = 1, shuffle_row_groups = False, workers_count = 2) as train_loader, \
       converter_test.make_torch_dataloader(batch_size = batch_size, transform_spec = get_transform_spec(),num_epochs = 1,shuffle_row_groups = False, workers_count = 2) as test_loader:
       for batch_idx, row_batch in enumerate(train_loader):
         break
         

row_batch.keys()

print(row_batch['hashed_wide'].shape)
print(row_batch['embedding_indexed'].shape)
print(row_batch['label'].shape)
print(row_batch['date'].shape)
print(row_batch['id'].shape)

print(row_batch['hashed_wide'].dtype) # float32, forced by vector Assembler
print(row_batch['embedding_indexed'].dtype)
print(row_batch['label'].dtype)
print(row_batch['date'].dtype)
print(row_batch['id'].dtype)

row_batch['label']

row_batch['hashed_wide'][:2,:]

row_batch['embedding_indexed'][0,:]

row_batch['date']
